timer fixed

+++ src/components/Timer.js
@@ -10,34 +10,37 @@ const Timer = () => {
   const [minutes, setMinutes] = useState(initialMinutes);
   const [seconds, setSeconds] = useState(initialSeconds);
 
+  const [quit, setQuit] = useState(false);
+
   useEffect(() => {
-    let interval = setInterval(() => {
-      if (seconds > 0) {
-        setSeconds(seconds - 1);
-      }
-      if (seconds === 0) {
-        if (minutes === 0) {
+    if (seconds !== 0 && minutes !== 0) {
+      let interval = setInterval(() => {
+        if (seconds >= 1) {
+          setSeconds(seconds - 1);
           clearInterval(interval);
-        } else {
-          setMinutes(minutes - 1);
-          setSeconds(59);
+          console.log(seconds);
         }
-      }
-      if (seconds === 0 || minutes === 0) {
-        localStorage.clear();
-        window.location = '/';
-      }
-    }, 1000);
+        if (seconds === 0) {
+          setSeconds(59);
+          clearInterval(interval);
+          if (minutes === 0) {
+            clearInterval(interval);
+          } else {
+            setMinutes(minutes - 1);
+            //   clearInterval(interval);
+          }
+        }
+        if (seconds === 0 && minutes === 0) {
+          localStorage.clear();
+          clearInterval(interval);
+          window.location = '/';
+        }
+      }, 1000);
+    } else {
+      setQuit(true);
+    }
   });
 
-  return (
-    <div>
-      {seconds < 0 || minutes < 0 ? (
-        <div>Time is over</div>
-      ) : (
-        `${minutes}:${seconds}`
-      )}
-    </div>
-  );
+  return <div>{quit ? <div>Time is over</div> : `${minutes}:${seconds}`}</div>;
 };
 export default Timer;
